requirements.txt
Flask
Flask-Cors
python-dotenv
sqlalchemy
passlib[bcrypt]
bcrypt
psycopg2-binary
